+++ tests/settings/resources/logging.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+LOGGING_INCLUDED = True