+++ tests/test_split_settings.py
@@ -7,6 +7,36 @@ def test_merge(merged):
     assert merged.STATIC_ROOT
 
 
+def test_alt_ext(alt_ext):
+    """Test that all values from settings are present."""
+    assert alt_ext.NO_EXT_INCLUDED
+    assert alt_ext.DOT_CONF_INCLUDED
+    assert alt_ext.DOUBLE_EXT_INCLUDED
+    assert alt_ext.OPTIONAL_INCLUDED
+
+
+def test_resources(resources):  # noqa: WPS218
+    """Test that all values from settings are present."""
+    assert resources.APPS_MIDDLEWARE_INCLUDED
+    assert resources.BASE_INCLUDED
+    assert resources.DATABASE_INCLUDED
+    assert resources.LOCALE_INCLUDED
+    assert resources.LOGGING_INCLUDED
+    assert resources.STATIC_SETTINGS_INCLUDED
+    assert resources.TEMPLATES_INCLUDED
+    assert resources.OVERRIDE_WORKS
+    assert resources.MULTIPLE_SUBDIR_FILE_INCLUDED
+    assert resources.UNFOUND_RESOURCE_IS_IOERROR
+    assert resources.UNFOUND_PACKAGE_IS_MODULE_ERROR
+    assert resources.UNFOUND_RESOURCE_ERR_STR.startswith(
+        'No such file',
+    )
+    assert resources.UNFOUND_RESOURCE_ERR_STR.endswith(
+        'does_not_exist.conf',
+    )
+    assert resources.UNFOUND_PACKAGE_ERR_STR == 'No module named does.not.exist'
+
+
 def test_override(merged, monkeypatch):
     """This setting must be overridden in the testing.py."""
     monkeypatch.setenv('DJANGO_SETTINGS_MODULE', 'tests.settings.merged')